refactor: . Move free function renderListItem() to TaskLineRenderer.ts
