+++ src/Renderer/TaskLineRenderer.ts
@@ -14,6 +14,7 @@ import { defaultTaskSaver } from '../ui/Menus/TaskEditingMenu';
 import { promptForDate } from '../ui/Menus/DatePicker';
 import { splitDateText } from '../DateTime/Postponer';
 import { DateMenu } from '../ui/Menus/DateMenu';
+import type { ListItem } from '../Task/ListItem';
 import { TaskFieldRenderer } from './TaskFieldRenderer';
 
 /**
@@ -402,4 +403,59 @@ export class TaskLineRenderer {
             });
         });
     }
+}
+
+export async function renderListItem(
+    taskList: HTMLUListElement,
+    listItem: ListItem,
+    listItemIndex: number,
+    textRenderer: any,
+    obsidianComponent: Component | null,
+) {
+    const li = createAndAppendElement('li', taskList);
+
+    if (listItem.statusCharacter) {
+        const checkbox = createAndAppendElement('input', li);
+        checkbox.classList.add('task-list-item-checkbox');
+        checkbox.type = 'checkbox';
+
+        checkbox.addEventListener('click', (event: MouseEvent) => {
+            event.preventDefault();
+            // It is required to stop propagation so that obsidian won't write the file with the
+            // checkbox (un)checked. Obsidian would write after us and overwrite our change.
+            event.stopPropagation();
+
+            // Should be re-rendered as enabled after update in file.
+            checkbox.disabled = true;
+
+            const checkedOrUncheckedListItem = listItem.checkOrUncheck();
+            replaceTaskWithTasks({ originalTask: listItem, newTasks: checkedOrUncheckedListItem });
+        });
+
+        if (listItem.statusCharacter !== ' ') {
+            checkbox.checked = true;
+            li.classList.add('is-checked');
+        }
+
+        li.classList.add('task-list-item');
+
+        // Set these to be compatible with stock obsidian lists:
+        li.setAttribute('data-task', listItem.statusCharacter.trim());
+        // Trim to ensure empty attribute for space. Same way as obsidian.
+        li.setAttribute('data-line', listItemIndex.toString());
+    }
+
+    const span = createAndAppendElement('span', li);
+    await textRenderer(listItem.description, span, listItem.findClosestParentTask()?.path ?? '', obsidianComponent);
+
+    // Unwrap the p-tag that was created by the MarkdownRenderer:
+    const pElement = span.querySelector('p');
+    if (pElement !== null) {
+        while (pElement.firstChild) {
+            span.insertBefore(pElement.firstChild, pElement);
+        }
+        pElement.remove();
+    }
+
+    return li;
 }
